+++ src/lib/components/PathField.svelte
@@ -20,7 +20,7 @@
 
 	<Button.Root
 		class="flex flex-grow px-3 py-1 items-center text-right rounded-lg group bg-gray-900 truncate
-            border border-gray-500 border-opacity-0 hover:border-opacity-100 relative"
+            border border-gray-500 border-opacity-0 hover:border-opacity-100"
 		on:click={onClick}
 	>
 		<div class="mr-2 rounded">